requirements.txt
# Core
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-telegram-bot>=21.0

# Tools
google-api-python-client>=2.100.0
google-auth>=2.23.0
duckduckgo_search>=6.0.0

# Utils
pyyaml>=6.0
httpx>=0.27.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
jinja2>=3.1.0
gitpython>=3.1.0

# Export / Publishing
pypandoc>=1.14
ebooklib>=0.18
Pillow>=10.0
weasyprint>=62.0

# ISBN / Barcode
isbnlib>=3.10
python-barcode>=0.15

# Web Preview
markdown>=3.4
Jinja2>=3.1

# MCP
fastmcp>=0.4
